feat: add shared financial data module
